refactor: move stream video to own plug
