+++ lib/media_server_web/helpers.ex
@@ -1,48 +1,4 @@
 defmodule MediaServerWeb.Helpers do
-  import Plug.Conn
-
-  def get_file_size(path) do
-    {:ok, %{size: size}} = File.stat(path)
-
-    size
-  end
-
-  def handle_range({"range", "bytes=0-1"}, conn, path) do
-    file_size = get_file_size(path)
-
-    conn
-    |> put_resp_header("content-type", "video/mp4")
-    |> put_resp_header("content-range", "bytes 0-1/#{file_size}")
-    |> send_file(206, path, 0, 2)
-  end
-
-  def handle_range({"range", "bytes=" <> start_pos}, conn, path) do
-    file_size = get_file_size(path)
-
-    offset =
-      String.split(start_pos, "-")
-      |> hd
-      |> String.to_integer()
-
-    conn
-    |> put_resp_header("content-type", "video/mp4")
-    |> put_resp_header("content-range", "bytes #{offset}-#{file_size - 1}/#{file_size}")
-    |> send_file(206, path, offset, file_size - offset)
-  end
-
-  def handle_range(nil, conn, path) do
-    file_size = get_file_size(path)
-
-    conn
-    |> put_resp_header("content-type", "video/mp4")
-    |> put_resp_header("content-range", "bytes 0-#{file_size - 1}/#{file_size}")
-    |> send_file(206, path, 0, file_size - 0)
-  end
-
-  def send_video(conn, headers, path) do
-    List.keyfind(headers, "range", 0)
-    |> handle_range(conn, path)
-  end
 
   def minutes_remaining_from_seconds(duration, current_time) do
     ceil((duration - current_time) / 60)
