+++ lib/media_server_web/router.ex
@@ -60,9 +60,9 @@ defmodule MediaServerWeb.Router do
     pipe_through :api
 
     get "/movies/:id/stream", StreamMovieController, :show
-    get "/movies/:id/subtitle", SubtitleMovieController, :show
-
     get "/episodes/:id/stream", StreamEpisodeController, :show
+
+    get "/movies/:id/subtitle", SubtitleMovieController, :show
     get "/episodes/:id/subtitle", SubtitleEpisodeController, :show
 
     post "/webhooks/movie", Webhooks.MovieController, :create
